requirements.txt
fastapi[standard]
pymongo 
python-dotenv
python-multipart
cloudinary
bcrypt 
pyjwt 